--- src/components/TitleSection/TitleSection.styled.tsx
@@ -1,14 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Title = styled.h2`
-  font-size: 24px;
-  color: #000000;
-  font-weight: 700;
-`;
-
-export const Description = styled.p`
-  font-size: 12px;
-  color: #0a0d13;
-  font-weight: 400;
-  margin: 12px 0px;
-`;
